+++ Mvc5Memberships/Areas/Admin/Controllers/SectionController.cs
@@ -112,8 +112,15 @@ namespace Mvc5Memberships.Areas.Admin.Controllers
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
             Section section = await db.Sections.FindAsync(id);
-            db.Sections.Remove(section);
-            await db.SaveChangesAsync();
+            var isUnused = await db.Items.CountAsync(i => i.SectionId == id) == 0;
+
+            if (isUnused)
+            {
+                db.Sections.Remove(section);
+                await db.SaveChangesAsync();
+            }
+
+            
             return RedirectToAction("Index");
         }
